MOVE ONLY: Move TapTree to taproot module
